Update/2194 replace connect account page exp (#2196)
* Remove experiment and replace the default experiment in favor of the treatment experience
* Update test snap
