--- client/connect-account-page-variant/style.scss
@@ -1,128 +1,0 @@
-.connect-account-page-variant {
-	.wcpay-connect-account-page-masthead {
-		position: relative;
-		height: 120px;
-		border-bottom: 1px solid $studio-gray-5;
-
-		h1 {
-			position: absolute;
-			top: 24px;
-			left: 24px;
-		}
-
-		.wcpay-connect-account-page-recommended {
-			position: absolute;
-			top: 44px;
-			left: 246px;
-		}
-
-		.wcpay-connect-account-page-hero {
-			@include breakpoint( '<660px' ) {
-				display: none;
-			}
-			position: absolute;
-			top: -100px;
-			right: -36px;
-			clip: rect( 100px, 220px, 220px, auto );
-		}
-	}
-
-	.wcpay-connect-account-page-pill {
-		/* adapted from https://github.com/woocommerce/woocommerce-admin/blob/main/packages/components/src/pill/style.scss */
-		font-size: 12px;
-		border: 1px solid $studio-gray-70;
-		border-radius: 28px;
-		color: $studio-gray-70;
-		display: inline-block;
-		padding: 2px 8px;
-	}
-
-	.wcpay-connect-account-page-payment-methods {
-		svg {
-			display: inline-block;
-			height: 35px;
-			width: 52px;
-			margin-right: 0.5em;
-			margin-bottom: 0.5em;
-		}
-	}
-
-	.wcpay-connect-account-page-terms-of-service {
-		display: block;
-		margin: 0.5em 0;
-	}
-
-	.connect-page-onboarding-steps {
-		@include breakpoint( '>660px' ) {
-			display: flex;
-			flex-direction: row;
-		}
-
-		.connect-page-onboarding-steps-item {
-			margin-top: 30px;
-
-			@include breakpoint( '>660px' ) {
-				margin-top: 0;
-				margin-right: 80px;
-			}
-		}
-	}
-
-	.wcpay-connect-account-page-step-number {
-		background-color: $studio-woocommerce-purple-50;
-		color: $white;
-		font-size: 20px;
-		border-radius: 50%;
-		min-height: 32px;
-		min-width: 32px;
-		display: inline-block;
-		text-align: center;
-		line-height: 32px;
-	}
-
-	.connect-account {
-		h2 {
-			font-size: 20px;
-			font-weight: 300;
-		}
-
-		h3 {
-			font-size: 1em;
-		}
-
-		.connect-account__card {
-			.woocommerce-card__body {
-				padding: 0;
-
-				.content {
-					padding: 0 24px 24px 24px;
-
-					a {
-						text-decoration: none;
-					}
-
-					hr {
-						position: relative;
-						left: -24px;
-						width: calc( 100% + 48px );
-					}
-				}
-			}
-		}
-
-		.connect-account__action {
-			margin-top: 16px;
-			margin-bottom: 0;
-		}
-
-		.connect-account__steps {
-			.woocommerce-card__body {
-				padding: 24px;
-			}
-
-			h2 {
-				margin-top: 0;
-			}
-		}
-	}
-}
