+++ client/connect-account-page/cards/cb.js
