+++ client/connect-account-page/cards/diners.js
