+++ client/connect-account-page/cards/discover.js
