+++ client/connect-account-page/cards/mastercard.js
