--- client/connect-account-page/hero-image.js
@@ -1,136 +1,0 @@
-/** @format */
-
-export default ( props ) => (
-	/* eslint-disable */
-	<svg
-		{ ...props }
-		width="260"
-		height="193"
-		viewBox="0 0 260 193"
-		fill="none"
-		xmlns="http://www.w3.org/2000/svg"
-	>
-		<g clipPath="url(#clip0)">
-			<path
-				d="M188.147 0.353976L5.49328 57.1468C1.37508 58.4273 -0.926025 62.806 0.353615 66.9269L33.2129 172.746C34.4926 176.867 38.8684 179.169 42.9866 177.889L225.64 121.096C229.758 119.815 232.059 115.437 230.78 111.316L197.92 5.49701C196.641 1.37611 192.265 -0.926507 188.147 0.353976Z"
-				fill="url(#paint0_linear)"
-			/>
-			<path
-				d="M187.201 2.02519L8.57824 57.5647C4.46004 58.8451 2.15894 63.2238 3.43858 67.3447L34.9709 168.89C36.2505 173.011 40.6264 175.314 44.7446 174.033L223.367 118.494C227.485 117.213 229.786 112.835 228.507 108.714L196.974 7.16823C195.695 3.04733 191.319 0.744715 187.201 2.02519Z"
-				fill="white"
-			/>
-			<path
-				d="M199.423 15.0552L5.8877 75.2317L11.0743 91.9346L204.61 31.7581L199.423 15.0552Z"
-				fill="#674399"
-			/>
-			<path
-				d="M95.0713 134.624L45.8979 149.914L48.1408 157.137L97.3142 141.847L95.0713 134.624Z"
-				fill="#BDBDBD"
-			/>
-			<path
-				d="M127.377 111.214L42.1133 137.725L44.3562 144.948L129.62 118.437L127.377 111.214Z"
-				fill="#E0E0E0"
-			/>
-			<path
-				d="M252.192 66.439H59.7576C55.4451 66.439 51.9492 69.9372 51.9492 74.2524V185.187C51.9492 189.502 55.4451 193 59.7576 193H252.192C256.504 193 260 189.502 260 185.187V74.2524C260 69.9372 256.504 66.439 252.192 66.439Z"
-				fill="url(#paint1_linear)"
-			/>
-			<path
-				d="M249.166 67.7293H62.1194C57.8069 67.7293 54.311 71.2275 54.311 75.5428V181.877C54.311 186.193 57.8069 189.691 62.1194 189.691H249.166C253.479 189.691 256.975 186.193 256.975 181.877V75.5428C256.975 71.2275 253.479 67.7293 249.166 67.7293Z"
-				fill="white"
-			/>
-			<path
-				d="M108.411 126.936H102.506V143.718H108.411V126.936Z"
-				fill="#E0E0E0"
-			/>
-			<path
-				d="M115.97 126.936H110.064V143.718H115.97V126.936Z"
-				fill="#E0E0E0"
-			/>
-			<path
-				d="M123.528 126.936H117.623V143.718H123.528V126.936Z"
-				fill="#E0E0E0"
-			/>
-			<path
-				d="M136.756 126.936H130.851V143.718H136.756V126.936Z"
-				fill="#E0E0E0"
-			/>
-			<path
-				d="M144.314 126.936H138.409V143.718H144.314V126.936Z"
-				fill="#E0E0E0"
-			/>
-			<path
-				d="M151.872 126.936H145.967V143.718H151.872V126.936Z"
-				fill="#E0E0E0"
-			/>
-			<path
-				d="M165.1 126.936H159.195V143.718H165.1V126.936Z"
-				fill="#E0E0E0"
-			/>
-			<path
-				d="M172.659 126.936H166.753V143.718H172.659V126.936Z"
-				fill="#E0E0E0"
-			/>
-			<path
-				d="M180.217 126.936H174.312V143.718H180.217V126.936Z"
-				fill="#E0E0E0"
-			/>
-			<path
-				d="M193.445 126.936H187.54V143.718H193.445V126.936Z"
-				fill="#E0E0E0"
-			/>
-			<path
-				d="M201.003 126.936H195.098V143.718H201.003V126.936Z"
-				fill="#E0E0E0"
-			/>
-			<path
-				d="M208.562 126.936H202.657V143.718H208.562V126.936Z"
-				fill="#E0E0E0"
-			/>
-			<path
-				d="M102.414 99.6817H67.2051V119.002H102.414V99.6817Z"
-				fill="#674399"
-			/>
-			<path
-				opacity="0.1"
-				d="M225.643 163.042H204.064V184.635H225.643V163.042Z"
-				fill="#674399"
-			/>
-			<path
-				opacity="0.1"
-				d="M236.433 163.042H214.854V184.635H236.433V163.042Z"
-				fill="#674399"
-			/>
-		</g>
-		<defs>
-			<linearGradient
-				id="paint0_linear"
-				x1="115.955"
-				y1="178.574"
-				x2="116.743"
-				y2="-0.324663"
-				gradientUnits="userSpaceOnUse"
-			>
-				<stop stopColor="#808080" stopOpacity="0.25" />
-				<stop offset="0.53514" stopColor="#808080" stopOpacity="0.12" />
-				<stop offset="1" stopColor="#808080" stopOpacity="0.1" />
-			</linearGradient>
-			<linearGradient
-				id="paint1_linear"
-				x1="114339"
-				y1="86036.4"
-				x2="114339"
-				y2="29661"
-				gradientUnits="userSpaceOnUse"
-			>
-				<stop stopColor="#808080" stopOpacity="0.25" />
-				<stop offset="0.53514" stopColor="#808080" stopOpacity="0.12" />
-				<stop offset="1" stopColor="#808080" stopOpacity="0.1" />
-			</linearGradient>
-			<clipPath id="clip0">
-				<rect width="260" height="193" fill="white" />
-			</clipPath>
-		</defs>
-	</svg>
-	/* eslint-enable */
-);
