+++ client/connect-account-page/strings.js
@@ -6,23 +6,59 @@ import { __ } from '@wordpress/i18n';
 import { __experimentalCreateInterpolateElement as createInterpolateElement } from 'wordpress-element';
 
 export default {
+	button: __( 'Finish Setup', 'woocommerce-payments' ),
+
 	heading: __( 'WooCommerce Payments', 'woocommerce-payments' ),
-	description: createInterpolateElement(
-		__(
-			'Accept credit card payments the easy way! <a>No set up fees. No monthly fees.</a>',
+
+	learnMore: __( 'Learn more', 'woocommerce-payments' ),
+
+	onboarding: {
+		heading: __(
+			'Finish setup to enable credit card payments',
 			'woocommerce-payments'
 		),
-		{
-			a: (
-				// eslint-disable-next-line jsx-a11y/anchor-has-content
-				<a
-					href="https://docs.woocommerce.com/document/payments/faq/fees/"
-					target="_blank"
-					rel="noopener noreferrer"
-				/>
-			),
-		}
+		description: __(
+			'With WooCommerce Payments, you can securely accept major cards, Apple Pay, and payments in over 100 currencies. Track cash flow and manage recurring revenue directly from your store’s dashboard - with no setup costs or monthly fees.',
+			'woocommerce-payments'
+		),
+	},
+
+	paymentMethodsHeading: __(
+		'Accepted payment methods',
+		'woocommerce-payments'
 	),
+
+	stepsHeading: __(
+		'You’re only steps away from getting paid',
+		'woocommerce-payments'
+	),
+
+	step1: {
+		heading: __(
+			'Create and connect your account',
+			'woocommerce-payments'
+		),
+		description: __(
+			'To ensure safe and secure transactions, a WordPress.com account is required.',
+			'woocommerce-payments'
+		),
+	},
+
+	step2: {
+		heading: __( 'Provide a few business details', 'woocommerce-payments' ),
+		description: __(
+			'Next we’ll ask you to verify your business and payment details to enable deposits.',
+			'woocommerce-payments'
+		),
+	},
+
+	step3: {
+		heading: __( 'Setup complete!', 'woocommerce-payments' ),
+		description: __(
+			'You’re ready to start using the features and benefits of WooCommerce Payments.'
+		),
+	},
+
 	terms: createInterpolateElement(
 		__(
 			'By clicking “Set up,” you agree to the <a>Terms of Service</a>',
@@ -39,11 +75,12 @@ export default {
 			),
 		}
 	),
+
 	onboardingDisabled: [
 		__(
 			"We've temporarily paused new account creation.",
-			'woocommmerce-payments'
+			'woocommerce-payments'
 		),
-		__( "We'll notify you when we resume!", 'woocommmerce-payments' ),
+		__( "We'll notify you when we resume!", 'woocommerce-payments' ),
 	],
 };
