+++ client/connect-account-page/style.scss
@@ -1,45 +1,132 @@
-.connect-account {
-	.connect-account__card {
-		.woocommerce-card__body {
-			line-height: 1.5;
-			p,
-			h2 {
-				text-align: center;
+.connect-account-page {
+	.wcpay-connect-account-page-masthead {
+		position: relative;
+		height: 120px;
+		border-bottom: 1px solid $studio-gray-5;
+
+		h1 {
+			position: absolute;
+			top: 24px;
+			left: 24px;
+		}
+
+		.wcpay-connect-account-page-recommended {
+			position: absolute;
+			top: 44px;
+			left: 246px;
+		}
+
+		.wcpay-connect-account-page-hero {
+			@include breakpoint( '<660px' ) {
+				display: none;
 			}
-			h2 {
-				font-size: 20px;
-				font-size: 1.2rem;
-				font-weight: 300;
-				margin: 16px 0 8px;
+			position: absolute;
+			top: -100px;
+			right: -36px;
+			clip: rect( 100px, 220px, 220px, auto );
+		}
+	}
+
+	.wcpay-connect-account-page-pill {
+		/* adapted from https://github.com/woocommerce/woocommerce-admin/blob/main/packages/components/src/pill/style.scss */
+		font-size: 12px;
+		border: 1px solid $studio-gray-70;
+		border-radius: 28px;
+		color: $studio-gray-70;
+		display: inline-block;
+		padding: 2px 8px;
+	}
+
+	.wcpay-connect-account-page-payment-methods {
+		svg {
+			display: inline-block;
+			height: 35px;
+			width: 52px;
+			margin-right: 0.5em;
+			margin-bottom: 0.5em;
+		}
+	}
+
+	.wcpay-connect-account-page-terms-of-service {
+		display: block;
+		margin: 0.5em 0;
+	}
+
+	.connect-page-onboarding-steps {
+		@include breakpoint( '>660px' ) {
+			display: flex;
+			flex-direction: row;
+		}
+
+		.connect-page-onboarding-steps-item {
+			margin-top: 30px;
+
+			@include breakpoint( '>660px' ) {
+				margin-top: 0;
+				margin-right: 80px;
 			}
 		}
+	}
 
-		.hero-image {
-			margin: 16px auto 24px;
-			display: block;
+	.wcpay-connect-account-page-step-number {
+		background-color: $studio-woocommerce-purple-50;
+		color: $white;
+		font-size: 20px;
+		border-radius: 50%;
+		min-height: 32px;
+		min-width: 32px;
+		display: inline-block;
+		text-align: center;
+		line-height: 32px;
+	}
+
+	.connect-account {
+		h2 {
+			font-size: 20px;
+			font-weight: 300;
 		}
 
-		.connect-account__description {
-			margin: 0 0 16px;
-			font-weight: 300;
-			font-size: 14px;
-			color: $studio-gray-60;
+		h3 {
+			font-size: 1em;
 		}
 
-		.connect-account__terms {
-			margin: 0 0 16px;
-			font-weight: 300;
-			font-size: 12px;
-			color: $studio-gray-60;
+		.connect-account__card {
+			.woocommerce-card__body {
+				padding: 0;
+				line-height: 1.5;
+				p,
+				h2 {
+					text-align: center;
+				}
+
+				.content {
+					padding: 0 24px 24px 24px;
+
+					a {
+						text-decoration: none;
+					}
+
+					hr {
+						position: relative;
+						left: -24px;
+						width: calc( 100% + 48px );
+					}
+				}
+			}
 		}
 
 		.connect-account__action {
 			margin-top: 16px;
+			margin-bottom: 0;
+		}
 
-			.is-button {
-				height: 32px;
-				line-height: 30px;
-				padding: 0 20px 2px;
+		.connect-account__steps {
+			.woocommerce-card__body {
+				padding: 24px;
+			}
+
+			h2 {
+				margin-top: 0;
 			}
 		}
 	}
